--- src/File/Path/AbstractProcessor.php
@@ -1,76 +1,0 @@
-<?php
-namespace Josegonzalez\Upload\File\Path;
-
-use Cake\ORM\Entity;
-use Cake\ORM\Table;
-use Cake\Utility\Hash;
-
-abstract class AbstractProcessor
-{
-    /**
-     * Table instance.
-     *
-     * @var \Cake\ORM\Table
-     */
-    protected $table;
-
-    /**
-     * Entity instance.
-     *
-     * @var \Cake\ORM\Entity
-     */
-    protected $entity;
-
-    /**
-     * Array of uploaded data for this field
-     *
-     * @var array
-     */
-    protected $data;
-
-    /**
-     * Name of field
-     *
-     * @var string
-     */
-    protected $field;
-
-    /**
-     * Settings for processing a path
-     *
-     * @var array
-     */
-    protected $settings;
-
-    /**
-     * Constructor.
-     *
-     * @param \Cake\ORM\Table  $table the instance managing the entity
-     * @param \Cake\ORM\Entity $entity the entity to construct a path for.
-     * @param array            $data the data being submitted for a save
-     * @param string           $field the field for which data will be saved
-     * @param array            $settings the settings for the current field
-     */
-    public function __construct(Table $table, Entity $entity, $data, $field, $settings)
-    {
-        $this->table = $table;
-        $this->entity = $entity;
-        $this->data = $data;
-        $this->field = $field;
-        $this->settings = $settings;
-    }
-
-    /**
-     * Returns the basepath for the current field/data combination
-     *
-     * @return string
-     */
-    abstract public function basepath();
-
-    /**
-     * Returns the filename for the current field/data combination
-     *
-     * @return string
-     */
-    abstract public function filename();
-}
